+++ json_database/exceptions.py
@@ -1,0 +1,6 @@
+class InvalidItemID(ValueError):
+    """ ItemID is invalid """
+
+
+class DatabaseNotCommitted(FileNotFoundError):
+    """ Database has not been saved in yet """